Module hotshot::demos::sdemo

this is a demo for sequencing consensus Sequencing consensus demo

This module provides an implementation of the HotShot suite of traits that implements a basic demonstration of sequencing consensus.

These implementations are useful in examples and integration testing, but are not suitable for production use.

Structs

GenesisAfterStartSnafu 🔒

SNAFU context selector for the SDemoError::GenesisAfterStart variant

GenesisCantHaveTransactionsSnafu 🔒

SNAFU context selector for the SDemoError::GenesisCantHaveTransactions variant

GenesisFailedSnafu 🔒

SNAFU context selector for the SDemoError::GenesisFailed variant

InvalidBlockSnafu 🔒

SNAFU context selector for the SDemoError::InvalidBlock variant

PreviousStateMismatchSnafu 🔒

SNAFU context selector for the SDemoError::PreviousStateMismatch variant

ReusedTxnSnafu 🔒

SNAFU context selector for the SDemoError::ReusedTxn variant

SDemoGenesisBlock

genesis block

SDemoNormalBlock

Any block after genesis

SDemoState

sequencing demo entry state

SDemoTransaction

The transaction for the sequencing demo

Enums

SDemoBlock

The block for the sequencing demo

SDemoError

The error type for the sequencing demo