Struct hotshot::HotShot

pub struct HotShot<CONSENSUS: ConsensusType, TYPES: NodeType, I: NodeImplementation<TYPES>> {
    pub(crate) inner: Arc<HotShotInner<TYPES, I>>,
    pub(crate) transactions: Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>,
    pub(crate) hotstuff: Arc<RwLock<Consensus<TYPES, I::Leaf>>>,
    pub(crate) member_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>,
    pub(crate) replica_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>,
    pub(crate) next_leader_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>,
    pub(crate) da_leader_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>,
    pub(crate) send_network_lookup: UnboundedSender<Option<TYPES::Time>>,
    pub(crate) recv_network_lookup: Arc<Mutex<UnboundedReceiver<Option<TYPES::Time>>>>,
    pub(crate) id: u64,
    pub(crate) _pd: PhantomData<CONSENSUS>,
}

Thread safe, shared view of a HotShot

Fields
inner: Arc<HotShotInner<TYPES, I>>

Handle to internal hotshot implementation

transactions: Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>

Transactions (this is shared btwn hotshot and Consensus)

hotstuff: Arc<RwLock<Consensus<TYPES, I::Leaf>>>

The hotstuff implementation

member_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>

for sending/recv-ing things with the DA member task

replica_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>

for sending/recv-ing things with the replica task

next_leader_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>

for sending/recv-ing things with the next leader task

da_leader_channel_map: Arc<RwLock<SendToTasks<TYPES, I>>>

for sending/recv-ing things to the da leader

send_network_lookup: UnboundedSender<Option<TYPES::Time>>

for sending messages to network lookup task

recv_network_lookup: Arc<Mutex<UnboundedReceiver<Option<TYPES::Time>>>>

for receiving messages in the network lookup task

id: u64

uid for instrumentation

_pd: PhantomData<CONSENSUS>

Phantom data for consensus type

Implementations

impl<TYPES: NodeType, I: NodeImplementation<TYPES>> HotShot<TYPES::ConsensusType, TYPES, I>

pub async fn new(
public_key: TYPES::SignatureKey,
private_key: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
nonce: u64,
config: HotShotConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>,
storage: I::Storage,
quorum_exchange: I::QuorumExchange,
committee_exchange: I::CommitteeExchange,
initializer: HotShotInitializer<TYPES, I::Leaf>,
metrics: Box<dyn Metrics>
) -> Result<Self, HotShotError<TYPES>>

Creates a new hotshot with the given configuration options and sets it up with the given genesis block

pub async fn timeout_view(
&self,
current_view: TYPES::Time,
send_replica: UnboundedSender<ProcessedConsensusMessage<TYPES, I>>,
send_next_leader: Option<UnboundedSender<ProcessedConsensusMessage<TYPES, I>>>
)

Marks a given view number as timed out. This should be called a fixed period after a round is started.

If the round has already ended then this function will essentially be a no-op. Otherwise run_round will return shortly after this function is called.

Panics

Panics if the current view is not in the channel map

pub async fn timeout_da_view(
&self,
current_view: TYPES::Time,
send_da_member: UnboundedSender<ProcessedConsensusMessage<TYPES, I>>,
send_da_leader: Option<UnboundedSender<ProcessedConsensusMessage<TYPES, I>>>
)

time out a da view

pub async fn publish_transaction_async(
&self,
transaction: TYPES::Transaction
) -> Result<(), HotShotError<TYPES>>

Publishes a transaction to the network

Errors

Will generate an error if an underlying network error occurs

pub async fn get_state(&self) -> <I::Leaf as LeafType>::StateCommitmentType

Returns a copy of the state

Panics

Panics if internal state for consensus is inconsistent

pub async fn get_decided_leaf(&self) -> I::Leaf

Returns a copy of the last decided leaf

Panics

Panics if internal state for consensus is inconsistent

pub async fn init(
public_key: TYPES::SignatureKey,
private_key: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
node_id: u64,
config: HotShotConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>,
storage: I::Storage,
quorum_exchange: I::QuorumExchange,
committee_exchange: I::CommitteeExchange,
initializer: HotShotInitializer<TYPES, I::Leaf>,
metrics: Box<dyn Metrics>
) -> Result<HotShotHandle<TYPES, I>, HotShotError<TYPES>>where
HotShot<TYPES::ConsensusType, TYPES, I>: ViewRunner<TYPES, I>,

Initializes a new hotshot and does the work of setting up all the background tasks

Assumes networking implementation is already primed.

Underlying HotShot instance starts out paused, and must be unpaused

Upon encountering an unrecoverable error, such as a failure to send to a broadcast channel, the HotShot instance will log the error and shut down.

Errors

Will return an error when the storage failed to insert the first QuorumCertificate

pub async fn send_broadcast_message(
&self,
kind: impl Into<MessageKind<TYPES, I>>
) -> Result<(), NetworkError>

Send a broadcast message.

This is an alias for hotshot.inner.networking.broadcast_message(msg.into()).

Errors

Will return any errors that the underlying broadcast_message can return.

pub async fn send_direct_message(
&self,
kind: impl Into<MessageKind<TYPES, I>>,
recipient: TYPES::SignatureKey
) -> Result<(), NetworkError>

Send a direct message to a given recipient.

This is an alias for hotshot.inner.networking.message_node(msg.into(), recipient).

Errors

Will return any errors that the underlying message_node can return.

pub(crate) async fn handle_broadcast_consensus_message(
&self,
msg: ConsensusMessage<TYPES, I>,
sender: TYPES::SignatureKey
)

Handle an incoming [ConsensusMessage] that was broadcasted on the network.

pub(crate) async fn handle_message(
&self,
item: Message<TYPES, I>,
transmit_type: TransmitType
)

decide which handler to call based on the message variant and transmit_type

pub(crate) async fn handle_direct_consensus_message(
&self,
msg: ConsensusMessage<TYPES, I>,
sender: TYPES::SignatureKey
)

Handle an incoming [ConsensusMessage] directed at this node.

pub(crate) async fn handle_broadcast_data_message(
&self,
msg: DataMessage<TYPES>,
_sender: TYPES::SignatureKey
)

Handle an incoming [DataMessage] that was broadcasted on the network

pub(crate) async fn handle_direct_data_message(
&self,
msg: DataMessage<TYPES>,
_sender: TYPES::SignatureKey
)

Handle an incoming [DataMessage] that directed at this node

pub fn get_next_view_timeout(&self) -> u64

return the timeout for a view for self

pub async fn create_or_obtain_chan_from_read(
view_num: TYPES::Time,
channel_map: RwLockUpgradableReadGuard<'_, SendToTasks<TYPES, I>>
) -> ViewQueue<TYPES, I>

given a view number and a upgradable read lock on a channel map, inserts entry into map if it doesn’t exist, or creates entry. Then returns a clone of the entry

pub async fn create_or_obtain_chan_from_write(
view_num: TYPES::Time,
channel_map: RwLockWriteGuard<'_, SendToTasks<TYPES, I>>
) -> ViewQueue<TYPES, I>

given a view number and a write lock on a channel map, inserts entry into map if it doesn’t exist, or creates entry. Then returns a clone of the entry