Module hotshot::tasks

Provides a number of tasks that run continuously

Modules

task_state

Provides trait to create task states from a SystemContextHandle

Enums

GlobalEvent

event for global event stream

Traits

EventTransformerState

Trait for intercepting and modifying messages between the network and consensus layers.

Functions

add_consensus_tasks

Adds consensus-related tasks to a SystemContextHandle.

add_network_event_task

Add the network task to handle events and send messages.

add_network_event_tasks

Adds the NetworkEventTaskState tasks.

add_network_message_and_request_receiver_tasks

Adds the NetworkMessageTaskState tasks and the request / receiver tasks.

add_network_message_task

Add the network task to handle messages and publish events.

add_network_tasks

adds tasks for sending/receiving messages to/from the network.

add_request_network_task

Add tasks for network requests and responses

add_response_task

Add a task which responds to requests on the network.

create_shutdown_event_monitor

Creates a monitor for shutdown events.