Trait hotshot::HotShotType

pub trait HotShotType<TYPES: NodeType, I: NodeImplementation<TYPES>> {
    // Required methods
    fn transactions(
        &self
    ) -> &Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>;
    fn hotstuff(&self) -> &Arc<RwLock<Consensus<TYPES, I::Leaf>>>;
    fn spawn_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = HotShotHandle<TYPES, I>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_broadcast_consensus_message<'life0, 'async_trait>(
        &'life0 self,
        msg: I::ConsensusMessage,
        sender: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_direct_consensus_message<'life0, 'async_trait>(
        &'life0 self,
        msg: I::ConsensusMessage,
        sender: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn handle_message<'life0, 'async_trait>(
        &'life0 self,
        item: Message<TYPES, I>,
        transmit_type: TransmitType
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn handle_broadcast_data_message<'life0, 'async_trait>(
        &'life0 self,
        msg: DataMessage<TYPES>,
        _sender: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn handle_direct_data_message<'life0, 'async_trait>(
        &'life0 self,
        msg: DataMessage<TYPES>,
        _sender: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

HotShot implementations that depend on [TYPES::ConsensusType].

Required Methods

fn transactions( &self ) -> &Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>

Get the [transactions] field of HotShot.

fn hotstuff(&self) -> &Arc<RwLock<Consensus<TYPES, I::Leaf>>>

Get the [hotstuff] field of HotShot.

fn spawn_all<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = HotShotHandle<TYPES, I>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Spawn all tasks that operate on the given HotShot.

For a list of which tasks are being spawned, see this module’s documentation.

fn handle_broadcast_consensus_message<'life0, 'async_trait>( &'life0 self, msg: I::ConsensusMessage, sender: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Handle an incoming [ConsensusMessage] that was broadcasted on the network.

fn handle_direct_consensus_message<'life0, 'async_trait>( &'life0 self, msg: I::ConsensusMessage, sender: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Handle an incoming [ConsensusMessage] directed at this node.

Provided Methods

fn handle_message<'life0, 'async_trait>( &'life0 self, item: Message<TYPES, I>, transmit_type: TransmitType ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

decide which handler to call based on the message variant and transmit_type

fn handle_broadcast_data_message<'life0, 'async_trait>( &'life0 self, msg: DataMessage<TYPES>, _sender: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Handle an incoming [DataMessage] that was broadcasted on the network

fn handle_direct_data_message<'life0, 'async_trait>( &'life0 self, msg: DataMessage<TYPES>, _sender: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Handle an incoming [DataMessage] that directed at this node