Trait hotshot::TwinsHandlerState

pub trait TwinsHandlerState<MESSAGE>
where
    Self: Sync,
    MESSAGE: Send + Sync,
{
    // Required methods
    fn new() -> Self;
    fn send_handler<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: &'life1 MESSAGE
    ) -> Pin<Box<dyn Future<Output = Either<MESSAGE, MESSAGE>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn recv_handler<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: &'life1 Either<MESSAGE, MESSAGE>
    ) -> Pin<Box<dyn Future<Output = MESSAGE> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn send_handler_arc<'life0, 'life1, 'async_trait>(
        lock: &'life0 Arc<RwLock<Self>>,
        event: &'life1 MESSAGE
    ) -> Pin<Box<dyn Future<Output = Either<MESSAGE, MESSAGE>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn recv_handler_arc<'life0, 'life1, 'async_trait>(
        lock: &'life0 Arc<RwLock<Self>>,
        event: &'life1 Either<MESSAGE, MESSAGE>
    ) -> Pin<Box<dyn Future<Output = MESSAGE> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Trait for handling messages for a node with a twin copy of consensus

Required Methods

fn new() -> Self

Initialize the state

fn send_handler<'life0, 'life1, 'async_trait>( &'life0 mut self, event: &'life1 MESSAGE ) -> Pin<Box<dyn Future<Output = Either<MESSAGE, MESSAGE>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handle a message sent to the twin from the outside, forwarding it to one of the two twins.

fn recv_handler<'life0, 'life1, 'async_trait>( &'life0 mut self, event: &'life1 Either<MESSAGE, MESSAGE> ) -> Pin<Box<dyn Future<Output = MESSAGE> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handle a message from either twin, forwarding it to the outside

Provided Methods

fn send_handler_arc<'life0, 'life1, 'async_trait>( lock: &'life0 Arc<RwLock<Self>>, event: &'life1 MESSAGE ) -> Pin<Box<dyn Future<Output = Either<MESSAGE, MESSAGE>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Wrapper for send_handler.

fn recv_handler_arc<'life0, 'life1, 'async_trait>( lock: &'life0 Arc<RwLock<Self>>, event: &'life1 Either<MESSAGE, MESSAGE> ) -> Pin<Box<dyn Future<Output = MESSAGE> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Wrapper for recv_handler.

Object Safety

This trait is not object safe.