Module hotshot::traits::election::vrf

generic vrf based election over anything that implements jf-primitives::vrf::Vrf

Structs

BinomialQuery

represents a binomial query made by sortition B(stake_attempt; replicas_stake; sortition_parameter / total_stake)

JfPubKey

type wrapper for VRF’s public key

VRFStakeTable

The stake table for VRFs

VRFStakeTableConfig

configuration specifying the stake table

VRFVoteToken

TODO doc me

VrfImpl

the vrf implementation

Constants

SORTITION_PARAMETER

the sortition committee size parameter

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

Functions

calculate_threshold 🔒

calculate_threshold_from_cache 🔒

check_bin_idx 🔒

checks that the expected aomunt of stake matches the VRF output TODO this can be optimized most likely

factorial 🔒

compute i! as a biguint

find_bin_idx 🔒

find the amount of stake we rolled. NOTE: in the future this requires a view numb Returns None if zero stake was rolled

generate_view_seed 🔒

generates the seed from algorand paper baseed on view_number and a constant as of now, but in the future will be other things this is a stop-gap

Type Definitions

BlsPubKey

concrete type for bls public key