Module hotshot::traits::implementations

Module for publicly usable implementations of the traits

Structs

CombinedNetworks

A communication channel with 2 networks, where we can fall back to the slower network if the primary fails

Libp2pNetwork

Networking implementation that uses libp2p generic over M which is the message type

MasterMap

Shared state for in-memory mock networking.

MemoryNetwork

In memory only network simulator.

NetworkingMetricsValue

Contains several NetworkingMetrics that we’re interested in from the networking interfaces

ProductionDef

The production run definition for the Push CDN. Uses the real protocols and a Redis discovery client.

PushCdnNetwork

A communication channel to the Push CDN, which is a collection of brokers and a marshal that helps organize them all. Is generic over both the type of key and the network protocol.

TestingDef

The testing run definition for the Push CDN. Uses the real protocols, but with an embedded discovery client.

UnderlyingCombinedNetworks

Wrapper for the tuple of WebServerNetwork and Libp2pNetwork We need this so we can impl TestableNetworkingImplementation on the tuple

WebServerNetwork

The web server network state

WrappedSignatureKey

A wrapped SignatureKey. We need to implement the Push CDN’s SignatureScheme trait in order to sign and verify messages to/from the CDN.

Type Aliases

PeerInfoVec

Type alias for a shared collection of peerid, multiaddrs