Module hotshot::traits::networking

Network access compatibility

This module contains a trait abstracting over network access, as well as implementations of that trait. Currently this includes

MemoryNetwork, an in memory testing-only implementation

Libp2pNetwork, a production-ready networking implementation built on top of libp2p-rs.

Modules

combined_network

Networking Implementation that has a primary and a fallback network. If the primary Errors we will use the backup to send or receive

libp2p_network

Libp2p based/production networking implementation This module provides a libp2p based networking implementation where each node in the network forms a tcp or udp connection to a subset of other nodes in the network

memory_network

In memory network simulator

push_cdn_network

The Push CDN network

web_server_network

A network implementation that connects to a web server.

Structs

FailedToSerializeSnafu

SNAFU context selector for the NetworkError::FailedToSerialize variant

InnerNetworkingMetrics

the set of counters and gauges for the networking metrics

NetworkingMetrics

The wrapper with a string name for the networking metrics

NetworkingMetricsValue

Contains several NetworkingMetrics that we’re interested in from the networking interfaces

Enums

NetworkError

Error type for networking

Constants

_DERIVE_core_fmt_Debug_FOR_InnerNetworkingMetrics 🔒

_DERIVE_core_fmt_Debug_FOR_NetworkingMetrics 🔒

_DERIVE_core_fmt_Debug_FOR_NetworkingMetricsValue 🔒

Traits

NetworkReliability

interface describing how reliable the network is