Enum hotshot::traits::networking::NetworkError

pub enum NetworkError {
    Libp2p {
        source: NetworkNodeHandleError,
    },
    MemoryNetwork {
        source: MemoryNetworkError,
    },
    CentralizedServer {
        source: CentralizedServerNetworkError,
    },
    UnimplementedFeature,
    CouldNotDeliver,
    NoSuchNode,
    FailedToSerialize {
        source: Box<ErrorKind, Global>,
    },
    FailedToDeserialize {
        source: Box<ErrorKind, Global>,
    },
    Timeout {
        source: TimeoutError,
    },
    ChannelSend,
    ShutDown,
    UnableToCancel,
}

Error type for networking

Variants

Libp2p

Fields

source: NetworkNodeHandleError

source of error

Libp2p specific errors

MemoryNetwork

Fields

source: MemoryNetworkError

source of error

memory network specific errors

CentralizedServer

Fields

source: CentralizedServerNetworkError

source of error

Centralized server specific errors

UnimplementedFeature

unimplemented functionality

CouldNotDeliver

Could not deliver a message to a specified recipient

NoSuchNode

Attempted to deliver a message to an unknown node

FailedToSerialize

Fields

source: Box<ErrorKind, Global>

Originating bincode error

Failed to serialize a network message

FailedToDeserialize

Fields

source: Box<ErrorKind, Global>

originating bincode error

Failed to deserealize a network message

Timeout

Fields

source: TimeoutError

Source of error

A timeout occurred

ChannelSend

Error sending output to consumer of NetworkingImplementation TODO this should have more information

ShutDown

The underlying connection has been shut down

UnableToCancel

unable to cancel a request, the request has already been cancelled