Struct hotshot::traits::networking::web_server_network::TaskMap

struct TaskMap<K: SignatureKey>(BTreeMap<u64, UnboundedSender<ConsensusIntentEvent<K>>>);

TaskMap is a wrapper around a BTreeMap that maps view numbers to tasks.

Each task is represented by a TaskChannel that can be used to send events to the task. The key K is a type that implements the SignatureKey trait.

Examples

let mut map: TaskMap<u64> = TaskMap::default();

Note

This struct is Clone, Deref, and DerefMut, so it can be used just like a BTreeMap.

Tuple Fields
0: BTreeMap<u64, UnboundedSender<ConsensusIntentEvent<K>>>
Implementations

impl<K: SignatureKey> TaskMap<K>

async fn prune_tasks( &mut self, current_view: u64, cancel_event_fn: fn(_: u64) -> ConsensusIntentEvent<K> )

Prunes tasks that are polling for a view less than or equal to current_view - 2.

This method cancels and removes all entries in the task map that are polling for a view less than or equal to current_view - 2. The cancellation is performed by sending a cancel_event to the task.

Arguments

current_view - The current view number. Tasks polling for a view less than or equal to current_view - 2 will be pruned.

cancel_event_fn - A function that takes a view number and returns a ConsensusIntentEvent to be sent to the task for cancellation.

Examples

let mut map: TaskMap<u64> = TaskMap::default();
map.prune_tasks(10, ConsensusIntentEvent::CancelPollForProposal(5)).await;