Trait hotshot::traits::node_implementation::NodeImplementation

pub trait NodeImplementation<TYPES>: Send + Sync + Clone + Eq + Hash + 'static + Serialize + for<'de> Deserialize<'de>
where
    TYPES: NodeType,
{
    type QuorumNetwork: ConnectedNetwork<Message<TYPES>, <TYPES as NodeType>::SignatureKey>;
    type CommitteeNetwork: ConnectedNetwork<Message<TYPES>, <TYPES as NodeType>::SignatureKey>;
    type Storage: Storage<TYPES>;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type QuorumNetwork: ConnectedNetwork<Message<TYPES>, <TYPES as NodeType>::SignatureKey>

Network for all nodes

type CommitteeNetwork: ConnectedNetwork<Message<TYPES>, <TYPES as NodeType>::SignatureKey>

Network for those in the DA committee

type Storage: Storage<TYPES>

Storage for DA layer interactions

Object Safety

This trait is not object safe.