Trait hotshot::traits::node_implementation::TestableNodeImplementation

pub trait TestableNodeImplementation<TYPES>: NodeImplementation<TYPES>where
    TYPES: NodeType,{
    type CommitteeElectionConfig;

    // Required methods
    fn committee_election_config_generator(
    ) -> Box<dyn Fn(u64) -> Self::CommitteeElectionConfig, Global>;
    fn state_create_random_transaction(
        state: Option<&<TYPES as NodeType>::StateType>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as BlockPayload>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Self::Leaf,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as BlockPayload>::Transaction;
    fn block_genesis() -> <TYPES as NodeType>::BlockType;
    fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Associated Types

type CommitteeElectionConfig

Election config for the DA committee

Required Methods

fn committee_election_config_generator( ) -> Box<dyn Fn(u64) -> Self::CommitteeElectionConfig, Global>

Generates a committee-specific election

fn state_create_random_transaction( state: Option<&<TYPES as NodeType>::StateType>, rng: &mut dyn RngCore, padding: u64 ) -> <<TYPES as NodeType>::BlockType as BlockPayload>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction( leaf: &Self::Leaf, rng: &mut dyn RngCore, padding: u64 ) -> <<TYPES as NodeType>::BlockType as BlockPayload>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> <TYPES as NodeType>::BlockType

generate a genesis block

fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Return the full internal state. This is useful for debugging.