Trait hotshot::traits::NetworkingImplementation

pub trait NetworkingImplementation<M, P>: Send + Sync where
    M: 'static + Serialize + DeserializeOwned + Send + Clone,
    P: 'static + SignatureKey, {
    fn ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn broadcast_message<'life0, 'async_trait>(
        &'life0 self, 
        message: M
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn message_node<'life0, 'async_trait>(
        &'life0 self, 
        message: M, 
        recipient: P
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn broadcast_queue<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn next_broadcast<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn direct_queue<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn next_direct<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn known_nodes<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Vec<P, Global>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn network_changes<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<NetworkChange<P>, Global>, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn shut_down<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn put_record<'life0, 'async_trait>(
        &'life0 self, 
        key: impl Serialize + Send + Sync + 'async_trait + 'static, 
        value: impl Serialize + Send + Sync + 'async_trait + 'static
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_record<'life0, 'async_trait, V>(
        &'life0 self, 
        key: impl Serialize + Send + Sync + 'async_trait + 'static
    ) -> Pin<Box<dyn Future<Output = Result<V, NetworkError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        V: 'async_trait + for<'a> Deserialize<'a>,
        Self: 'async_trait;
}

Describes, generically, the behaviors a networking implementation must have

Required Methods

fn ready<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns true when node is successfully initialized into the network

Blocks until node is ready

fn broadcast_message<'life0, 'async_trait>(
&'life0 self,
message: M
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Broadcasts a message to the network

Should provide that the message eventually reach all non-faulty nodes

fn message_node<'life0, 'async_trait>(
&'life0 self,
message: M,
recipient: P
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Sends a direct message to a specific node

fn broadcast_queue<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Moves out the entire queue of received broadcast messages, should there be any

Provided as a future to allow the backend to do async locking

fn next_broadcast<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Provides a future for the next received broadcast

Will unwrap the underlying NetworkMessage

fn direct_queue<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<M, Global>, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Moves out the entire queue of received direct messages to this node

fn next_direct<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<M, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Provides a future for the next received direct message to this node

Will unwrap the underlying NetworkMessage

fn known_nodes<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Vec<P, Global>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Node’s currently known to the networking implementation

Kludge function to work around leader election

fn network_changes<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<NetworkChange<P>, Global>, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a list of changes in the network that have been observed. Calling this function will clear the internal list.

fn shut_down<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Shut down this network. Afterwards this network should no longer be used.

This should also cause other functions to immediately return with a NetworkError

fn put_record<'life0, 'async_trait>(
&'life0 self,
key: impl Serialize + Send + Sync + 'async_trait + 'static,
value: impl Serialize + Send + Sync + 'async_trait + 'static
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Insert value into the shared store under key.

fn get_record<'life0, 'async_trait, V>(
&'life0 self,
key: impl Serialize + Send + Sync + 'async_trait + 'static
) -> Pin<Box<dyn Future<Output = Result<V, NetworkError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
V: 'async_trait + for<'a> Deserialize<'a>,
Self: 'async_trait,

Get value stored in shared store under key