Trait hotshot::traits::NodeImplementation

pub trait NodeImplementation<TYPES>: 'static + Send + Sync + Debug + Clonewhere
    TYPES: NodeType,{
    type Leaf: LeafType<NodeType = TYPES>;
    type Storage: Storage<TYPES, Self::Leaf> + Clone;
    type Membership: Membership<TYPES> + Debug;
    type Proposal: ProposalType<NodeType = TYPES>;
    type Vote: VoteType<TYPES>;
    type DAVoteAccumulator: Accumulator<(Commitment<<TYPES as NodeType>::BlockType>, (EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType))), BTreeMap<EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType), Global>>;
    type QuorumVoteAccumulator: Accumulator<(Commitment<Self::Leaf>, (EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType))), BTreeMap<EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType), Global>>;
    type Networking: CommunicationChannel<TYPES, Self::Proposal, Self::Vote, Self::Membership>;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type Leaf: LeafType<NodeType = TYPES>

type Storage: Storage<TYPES, Self::Leaf> + Clone

Storage type for this consensus implementation

type Membership: Membership<TYPES> + Debug

Membership Time is generic here to allow multiple implementations of membership trait for difference consensus protocols

type Proposal: ProposalType<NodeType = TYPES>

type Vote: VoteType<TYPES>

type DAVoteAccumulator: Accumulator<(Commitment<<TYPES as NodeType>::BlockType>, (EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType))), BTreeMap<EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType), Global>>

type QuorumVoteAccumulator: Accumulator<(Commitment<Self::Leaf>, (EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType))), BTreeMap<EncodedPublicKey, (EncodedSignature, <TYPES as NodeType>::VoteTokenType), Global>>

type Networking: CommunicationChannel<TYPES, Self::Proposal, Self::Vote, Self::Membership>

Networking type for this consensus implementation