Trait hotshot::traits::State

pub trait State: Serialize + DeserializeOwned + Clone + Debug + Default + Hash + PartialEq<Self> + Eq + Send + Sync + Committable {
    type Error: Error + Debug + Send + Sync;
    type BlockHeader: BlockHeader;
    type BlockPayload: BlockPayload;
    type Time: ConsensusTime;

    // Required methods
    fn validate_block(
        &self,
        block_header: &Self::BlockHeader,
        view_number: &Self::Time
    ) -> bool;
    fn initialize() -> Self;
    fn append(
        &self,
        block_header: &Self::BlockHeader,
        view_number: &Self::Time
    ) -> Result<Self, Self::Error>;
    fn on_commit(&self);
}

Abstraction over the state that blocks modify

This trait represents the behaviors that the ‘global’ ledger state must have:

A defined error type (Error)

The type of block that modifies this type of state (BlockPayload)

The ability to validate that a block is actually a valid extension of this state (validate_block)

The ability to produce a new state, with the modifications from the block applied (append)

Required Associated Types

type Error: Error + Debug + Send + Sync

The error type for this particular type of ledger state

type BlockHeader: BlockHeader

The type of block header this state is associated with

type BlockPayload: BlockPayload

The type of block payload this state is associated with

type Time: ConsensusTime

Time compatibility needed for reward collection

Required Methods

fn validate_block( &self, block_header: &Self::BlockHeader, view_number: &Self::Time ) -> bool

Returns true if and only if the provided block header is valid and can extend this state

fn initialize() -> Self

Initialize the state.

fn append( &self, block_header: &Self::BlockHeader, view_number: &Self::Time ) -> Result<Self, Self::Error>

Appends the given block header to this state, returning an new state

Errors

Should produce and error if appending this block header would lead to an invalid state

fn on_commit(&self)

Gets called to notify the persistence backend that this state has been committed