Trait hotshot::traits::TestableNodeImplementation

pub trait TestableNodeImplementation<CONSENSUS, TYPES>: NodeImplementation<TYPES>where
    CONSENSUS: ConsensusType,
    TYPES: NodeType<ConsensusType = CONSENSUS>,{
    type CommitteeCommChannel;
    type CommitteeNetwork;

    // Required methods
    fn network_generator(
        expected_node_count: usize,
        num_bootstrap: usize,
        da_committee_size: usize
    ) -> Box<dyn Fn(u64) -> <<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK + 'static, Global>
       where <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking: CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>;
    fn quorum_comm_channel_generator(
    ) -> Box<dyn Fn(Arc<<<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>
       where <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking: CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>;
    fn committee_comm_channel_generator(
    ) -> Box<dyn Fn(Arc<<<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> Self::CommitteeCommChannel + 'static, Global>;
    fn state_create_random_transaction(
        state: Option<&<TYPES as NodeType>::StateType>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Self::Leaf,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction;
    fn block_genesis() -> <TYPES as NodeType>::BlockType;
    fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn generate_test_key(
        id: u64
    ) -> <<TYPES as NodeType>::SignatureKey as SignatureKey>::PrivateKey;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Associated Types

type CommitteeCommChannel

Communication channel the DA committee. Only needed for the sequencing consensus.

type CommitteeNetwork

Connected network for the DA committee. Only needed for the sequencing consensus.

Required Methods

fn network_generator( expected_node_count: usize, num_bootstrap: usize, da_committee_size: usize ) -> Box<dyn Fn(u64) -> <<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK + 'static, Global>where <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking: CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>,

Generate a quorum network given an expected node count.

fn quorum_comm_channel_generator( ) -> Box<dyn Fn(Arc<<<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>where <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking: CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>,

Generate a quorum communication channel given the network.

fn committee_comm_channel_generator( ) -> Box<dyn Fn(Arc<<<<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <<Self::Exchanges as ExchangesType<<TYPES as NodeType>::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>>::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> Self::CommitteeCommChannel + 'static, Global>

Generate a committee communication channel given the network.

fn state_create_random_transaction( state: Option<&<TYPES as NodeType>::StateType>, rng: &mut dyn RngCore, padding: u64 ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction( leaf: &Self::Leaf, rng: &mut dyn RngCore, padding: u64 ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> <TYPES as NodeType>::BlockType

generate a genesis block

fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, Self: 'async_trait,

Return the full internal state. This is useful for debugging.

fn generate_test_key( id: u64 ) -> <<TYPES as NodeType>::SignatureKey as SignatureKey>::PrivateKey

The private key of the node id in a test.