Trait hotshot::traits::ValidatedState

pub trait ValidatedState: Serialize + DeserializeOwned + Clone + Debug + Default + Hash + PartialEq + Eq + Send + Sync {
    type Error: Error + Debug + Send + Sync;
    type Instance: InstanceState;
    type BlockHeader: BlockHeader<State = Self>;
    type BlockPayload: BlockPayload;
    type Time: ConsensusTime;

    // Required methods
    fn validate_and_apply_header(
        &self,
        instance: &Self::Instance,
        parent_header: &Self::BlockHeader,
        proposed_header: &Self::BlockHeader
    ) -> Result<Self, Self::Error>;
    fn from_header(block_header: &Self::BlockHeader) -> Self;
    fn on_commit(&self);

    // Provided method
    fn genesis(instance: &Self::Instance) -> Self { ... }
}

Abstraction over the state that blocks modify

This trait represents the behaviors that the ‘global’ ledger state must have:

A defined error type (Error)

The type of block that modifies this type of state (BlockPayload(ValidatedStates:: BlockPayload))

The ability to validate that a block header is actually a valid extension of this state and produce a new state, with the modifications from the block applied (validate_and_apply_header)

Required Associated Types

type Error: Error + Debug + Send + Sync

The error type for this particular type of ledger state

type Instance: InstanceState

The type of the instance-level state this state is assocaited with

type BlockHeader: BlockHeader<State = Self>

The type of block header this state is associated with

type BlockPayload: BlockPayload

The type of block payload this state is associated with

type Time: ConsensusTime

Time compatibility needed for reward collection

Required Methods

fn validate_and_apply_header( &self, instance: &Self::Instance, parent_header: &Self::BlockHeader, proposed_header: &Self::BlockHeader ) -> Result<Self, Self::Error>

Check if the proposed block header is valid and apply it to the state if so.

Returns the new state.

Arguments

instance - Immutable instance-level state.

Errors

If the block header is invalid or appending it would lead to an invalid state.

fn from_header(block_header: &Self::BlockHeader) -> Self

Construct the state with the given block header.

This can also be used to rebuild the state for catchup.

fn on_commit(&self)

Gets called to notify the persistence backend that this state has been committed

Provided Methods

fn genesis(instance: &Self::Instance) -> Self

Construct a genesis validated state.

Object Safety

This trait is not object safe.