Struct hotshot::types::SystemContextHandle

pub struct SystemContextHandle<TYPES: NodeType, I: NodeImplementation<TYPES>, V: Versions> {
    pub(crate) output_event_stream: (Sender<Event<TYPES>>, InactiveReceiver<Event<TYPES>>),
    pub(crate) internal_event_stream: (Sender<Arc<HotShotEvent<TYPES>>>, InactiveReceiver<Arc<HotShotEvent<TYPES>>>),
    pub(crate) consensus_registry: ConsensusTaskRegistry<HotShotEvent<TYPES>>,
    pub(crate) network_registry: NetworkTaskRegistry,
    pub hotshot: Arc<SystemContext<TYPES, I, V>>,
    pub(crate) storage: Arc<RwLock<I::Storage>>,
    pub network: Arc<I::Network>,
    pub memberships: Arc<RwLock<TYPES::Membership>>,
    pub epoch_height: u64,
}

Event streaming handle for a SystemContext instance running in the background

This type provides the means to message and interact with a background SystemContext instance, allowing the ability to receive Events from it, send transactions to it, and interact with the underlying storage.

Fields
output_event_stream: (Sender<Event<TYPES>>, InactiveReceiver<Event<TYPES>>)

The sender and receiver, to allow the application to communicate with HotShot.

internal_event_stream: (Sender<Arc<HotShotEvent<TYPES>>>, InactiveReceiver<Arc<HotShotEvent<TYPES>>>)

access to the internal event stream, in case we need to, say, shut something down

consensus_registry: ConsensusTaskRegistry<HotShotEvent<TYPES>>

registry for controlling consensus tasks

network_registry: NetworkTaskRegistry

registry for controlling network tasks

hotshot: Arc<SystemContext<TYPES, I, V>>

Internal reference to the underlying SystemContext

storage: Arc<RwLock<I::Storage>>

Reference to the internal storage for consensus datum.

network: Arc<I::Network>

Networks used by the instance of hotshot

memberships: Arc<RwLock<TYPES::Membership>>

Memberships used by consensus

epoch_height: u64

Number of blocks in an epoch, zero means there are no epochs

Implementations

impl<TYPES: NodeType, I: NodeImplementation<TYPES> + 'static, V: Versions> SystemContextHandle<TYPES, I, V>

pub fn add_task<S: TaskState<Event = HotShotEvent<TYPES>> + 'static>( &mut self, task_state: S, )

Adds a hotshot consensus-related task to the SystemContextHandle.

pub fn event_stream(&self) -> impl Stream<Item = Event<TYPES>>

obtains a stream to expose to the user

pub async fn send_external_message( &self, msg: Vec<u8>, recipients: RecipientList<TYPES::SignatureKey>, ) -> Result<()>

Message other participants with a serialized message from the application Receivers of this message will get an Event::ExternalMessageReceived via the event stream.

Errors

Errors if serializing the request fails, or the request fails to be sent

pub fn request_proposal( &self, view: TYPES::View, leaf_commitment: Commitment<Leaf2<TYPES>>, ) -> Result<impl Future<Output = Result<Proposal<TYPES, QuorumProposalWrapper<TYPES>>>>>

Request a proposal from the all other nodes. Will block until some node returns a valid proposal with the requested commitment. If nobody has the proposal this will block forever

Errors

Errors if signing the request for proposal fails

pub fn event_stream_known_impl(&self) -> Receiver<Event<TYPES>>

HACK so we can know the types when running tests… there are two cleaner solutions:

make the stream generic and in nodetypes or nodeimpelmentation

type wrapper

pub fn internal_event_stream_sender(&self) -> Sender<Arc<HotShotEvent<TYPES>>>

HACK so we can create dependency tasks when running tests

pub fn internal_event_stream_receiver_known_impl( &self, ) -> Receiver<Arc<HotShotEvent<TYPES>>>

HACK so we can know the types when running tests… there are two cleaner solutions:

make the stream generic and in nodetypes or nodeimpelmentation

type wrapper

NOTE: this is only used for sanity checks in our tests

pub async fn decided_state(&self) -> Arc<TYPES::ValidatedState>

Get the last decided validated state of the SystemContext instance.

Panics

If the internal consensus is in an inconsistent state.

pub async fn state( &self, view: TYPES::View, ) -> Option<Arc<TYPES::ValidatedState>>

Get the validated state from a given view.

Returns the requested state, if the SystemContext is tracking this view. Consensus tracks views that have not yet been decided but could be in the future. This function may return None if the requested view has already been decided (but see decided_state) or if there is no path for the requested view to ever be decided.

pub async fn decided_leaf(&self) -> Leaf2<TYPES>

Get the last decided leaf of the SystemContext instance.

Panics

If the internal consensus is in an inconsistent state.

pub fn try_decided_leaf(&self) -> Option<Leaf2<TYPES>>

Tries to get the most recent decided leaf, returning instantly if we can’t acquire the lock.

Panics

Panics if internal consensus is in an inconsistent state.

pub async fn submit_transaction( &self, tx: TYPES::Transaction, ) -> Result<(), HotShotError<TYPES>>

Submits a transaction to the backing SystemContext instance.

The current node broadcasts the transaction to all nodes on the network.

Errors

Will return a HotShotError if some error occurs in the underlying SystemContext instance.

pub fn consensus(&self) -> Arc<RwLock<Consensus<TYPES>>>

Get the underlying consensus state for this SystemContext

pub async fn shut_down(&mut self)

Shut down the inner hotshot and wait until all background threads are closed.

pub fn next_view_timeout(&self) -> u64

return the timeout for a view of the underlying SystemContext

pub async fn leader( &self, view_number: TYPES::View, epoch_number: Option<TYPES::Epoch>, ) -> Result<TYPES::SignatureKey>

Wrapper for HotShotConsensusApi’s leader function

Errors

Returns an error if the leader cannot be calculated

pub fn public_key(&self) -> TYPES::SignatureKey

Wrapper to get this node’s public key

pub fn external_channel_sender(&self) -> Sender<Event<TYPES>>

Get the sender side of the external event stream for testing purpose

pub fn internal_channel_sender(&self) -> Sender<Arc<HotShotEvent<TYPES>>>

Get the sender side of the internal event stream for testing purpose

pub async fn cur_view(&self) -> TYPES::View

Wrapper to get the view number this node is on.

pub async fn cur_epoch(&self) -> Option<TYPES::Epoch>

Wrapper to get the epoch number this node is on.

pub fn storage(&self) -> Arc<RwLock<I::Storage>>

Provides a reference to the underlying storage for this SystemContext, allowing access to historical data