Crate hotshot

Provides a generic rust implementation of the HotShot BFT protocol

Re-exports

pub use rand;

Modules

documentation

helpers

Contains helper functions for the crate

tasks

Provides a number of tasks that run continuously

traits

Contains traits consumed by SystemContext

types

Contains types used by the crate

Structs

HotShotInitializer

initializer struct for creating starting block

InitializerEpochInfo

SystemContext

Holds the state needed to participate in HotShot consensus

Enums

HotShotError

Reexport error type Error type for HotShot

Constants

H_256

Length, in bytes, of a 256 bit hash

H_512

Length, in bytes, of a 512 bit hash

Functions

load_start_epoch_info 🔒

Type Aliases

InitializerAnchor

The anchor leaf with its validated state and optional state delta, as supplied to HotShotInitializer::load. The state and delta may be omitted, in which case a sparse state is constructed from the leaf’s header.