Crate hotshot

Provides a generic rust implementation of the HotShot BFT protocol

See the protocol documentation for a protocol description.

Re-exports

pub use rand;

Modules

committee

Contains structures and functions for committee election

data

Provides types useful for representing HotShot’s data structures

demos

Contains implementations of the HotShot traits used in the examples and integration testing.

documentation

HotShot: A linear time, committee electing, BFT Protocol.

traits

Contains traits consumed by HotShot

types

Contains types used by the crate

Structs

HotShot

Thread safe, shared view of a HotShot

HotShotInner

Holds the state needed to participate in HotShot consensus

Enums

HotShotError

Reexport error type Error type for HotShot

Constants

H_256

Length, in bytes, of a 256 bit hash

H_512

Length, in bytes, of a 512 bit hash

Functions

create_or_obtain_chan_from_read

given a view number and a upgradable read lock on a channel map, inserts entry into map if it doesn’t exist, or creates entry. Then returns a clone of the entry

create_or_obtain_chan_from_write

given a view number and a write lock on a channel map, inserts entry into map if it doesn’t exist, or creates entry. Then returns a clone of the entry