Crate hotshot

Provides a generic rust implementation of the HotShot BFT protocol

See the protocol documentation for a protocol description.

Re-exports

pub use rand;

Modules

certificate

Data availability support Contains structures and functions for committee election

demos

Contains implementations of the HotShot traits used in the examples and integration testing.

documentation

HotShot: A linear time, committee electing, BFT Protocol.

tasks

Provides a number of tasks that run continuously on a HotShot

traits

Contains traits consumed by HotShot

types

Contains types used by the crate

Structs

HotShot

Thread safe, shared view of a HotShot

HotShotConsensusApi 🔒

A handle that exposes the interface that hotstuff needs to interact with HotShot

HotShotInitializer

initializer struct for creating starting block

HotShotInner

Holds the state needed to participate in HotShot consensus

Enums

HotShotError

Reexport error type Error type for HotShot

Constants

H_256

Length, in bytes, of a 256 bit hash

H_512

Length, in bytes, of a 512 bit hash

Traits

ViewRunner

A view runner implemented by HotShot for different types of consensus.