Crate hotshot

Provides a generic rust implementation of the HotShot BFT protocol

Re-exports

pub use rand;

Modules

documentation

helpers

Contains helper functions for the crate

tasks

Provides a number of tasks that run continuously

traits

Contains traits consumed by SystemContext

types

Contains types used by the crate

Structs

DoubleTwinsHandler

A TwinsHandlerState that forwards each message to both twins, and returns messages from each of them.

HotShotInitializer

initializer struct for creating starting block

InitializerEpochInfo

MarketplaceConfig

Wrapper for all marketplace config that needs to be passed when creating a new instance of HotShot

RandomTwinsHandler

A TwinsHandlerState that randomly forwards a message to either twin, and returns messages from both.

SystemContext

Holds the state needed to participate in HotShot consensus

Enums

HotShotError

Reexport error type Error type for HotShot

Constants

H_256

Length, in bytes, of a 256 bit hash

H_512

Length, in bytes, of a 512 bit hash

Traits

TwinsHandlerState

Trait for handling messages for a node with a twin copy of consensus

Functions

load_start_epoch_info ๐

Type Aliases

Channel ๐

An async broadcast channel