Module hotshot::tasks

Provides a number of tasks that run continuously

Modules

task_state

Provides trait to create task states from a SystemContextHandle

Enums

GlobalEvent

event for global event stream

Functions

add_consensus_task

add the consensus task

add_da_task

add the Data Availability task

add_network_event_task

Add the network task to handle events and send messages.

add_network_message_task

Add the network task to handle messages and publish events.

add_quorum_proposal_task

add the quorum proposal task

add_request_network_task

Add tasks for network requests and responses

add_response_task

Add a task which responds to requests on the network.

add_transaction_task

add the Transaction Handling task

add_upgrade_task

add the Upgrade task.

add_vid_task

add the VID task

add_view_sync_task

add the view sync task

inject_consensus_polls

Setup polls for the given consensus_state

inject_quorum_proposal_polls

Setup polls for the given quorum_proposal.