Trait hotshot::HotShotType

pub trait HotShotType<TYPES: NodeType, I: NodeImplementation<TYPES>> {
    // Required methods
    fn transactions(
        &self
    ) -> &Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>;
    fn consensus(&self) -> &Arc<RwLock<Consensus<TYPES, I::Leaf>>>;
    fn run_tasks<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, I>> + Send + 'async_trait>>
       where Self: 'async_trait;
}

[HotShot] implementations that depend on [TYPES::ConsensusType].

Required Methods

fn transactions( &self ) -> &Arc<SubscribableRwLock<HashMap<Commitment<TYPES::Transaction>, TYPES::Transaction>>>

Get the [transactions] field of [HotShot].

fn consensus(&self) -> &Arc<RwLock<Consensus<TYPES, I::Leaf>>>

Get the [hotstuff] field of [HotShot].

fn run_tasks<'async_trait>( self ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, I>> + Send + 'async_trait>>where Self: 'async_trait,

Spawn all tasks that operate on the given [HotShot].

For a list of which tasks are being spawned, see this module’s documentation.