Module hotshot::traits::implementations

Module for publicly usable implementations of the traits

Structs

CentralizedServerNetwork

Handle for connecting to a centralized server

DummyReliability

dummy implementation of network reliability

Libp2pNetwork

Networking implementation that uses libp2p generic over M which is the message type

MasterMap

Shared state for in-memory mock networking.

MemoryNetwork

In memory only network simulator.

MemoryStorage

In memory, ephemeral, storage for a HotShot instance

Stateless

Dummy, do nothing implementation of StatefulHandler

WNetwork

Handle to the underlying networking implementation

Type Definitions

PeerInfoVec

Type alias for a shared collection of peerid, multiaddrs