hotshot::traits
Module networking

Network access compatibility

This module contains a trait abstracting over network access, as well as implementations of that trait. Currently this includes

MemoryNetwork, an in memory testing-only implementation

Libp2pNetwork, a production-ready networking implementation built on top of libp2p-rs.

Modules

combined_network

Networking Implementation that has a primary and a fallback network. If the primary Errors we will use the backup to send or receive

libp2p_network

Libp2p based/production networking implementation This module provides a libp2p based networking implementation where each node in the network forms a tcp or udp connection to a subset of other nodes in the network

memory_network

In memory network simulator

push_cdn_network

The Push CDN network

Enums

NetworkError

Errors that can occur in the network

Traits

NetworkReliability

interface describing how reliable the network is