Module hotshot::traits::networking

Network access compatibility

This module contains a trait abstracting over network access, as well as implementations of that trait. Currently this includes

MemoryNetwork, an in memory testing-only implementation

Libp2pNetwork, a production-ready networking impelmentation built on top of libp2p-rs.

Modules

centralized_server_network

A network implementation that attempts to connect to a centralized server.

libp2p_network

Libp2p based/production networking implementation This module provides a libp2p based networking implementation where each node in the network forms a tcp or udp connection to a subset of other nodes in the network

memory_network

In memory network simulator

web_server_network

A network implementation that connects to a web server.

Structs

ChannelSendSnafu

SNAFU context selector for the NetworkError::ChannelSend variant

CouldNotDeliverSnafu

SNAFU context selector for the NetworkError::CouldNotDeliver variant

FailedToDeserializeSnafu

SNAFU context selector for the NetworkError::FailedToDeserialize variant

FailedToSerializeSnafu

SNAFU context selector for the NetworkError::FailedToSerialize variant

NetworkingMetrics 🔒

Contains the metrics that we’re interested in from the networking interfaces

NoSuchNodeSnafu

SNAFU context selector for the NetworkError::NoSuchNode variant

ShutDownSnafu

SNAFU context selector for the NetworkError::ShutDown variant

Enums

NetworkError

Error type for networking

Traits

NetworkReliability

interface describing how reliable the network is