Module hotshot::traits::networking::push_cdn_network

The Push CDN network

Structs

BrokerDef

The broker definition for the Push CDN. Uses the TCP protocol and trusted middleware.

CdnMetricsValue

CDN-specific metrics

ClientDef

The client definition for the Push CDN. Uses the TCP+TLS protocol and no middleware. Differs from the user definition in that is on the client-side.

KeyPair

We encapsulate keys here to help readability.

ProductionDef

The production run definition for the Push CDN. Uses the real protocols and a Redis discovery client.

PushCdnNetwork

A communication channel to the Push CDN, which is a collection of brokers and a marshal that helps organize them all. Is generic over both the type of key and the network protocol.

TestingDef

The testing run definition for the Push CDN. Uses the real protocols, but with an embedded discovery client.

UserDef

The user definition for the Push CDN. Uses the TCP+TLS protocol and untrusted middleware.

WrappedSignatureKey

A wrapped SignatureKey. We need to implement the Push CDN’s SignatureScheme trait in order to sign and verify messages to/from the CDN.

Enums

Topic

The enum for the topics we can subscribe to in the Push CDN