Trait hotshot::traits::NodeImplementation

pub trait NodeImplementation: 'static + Send + Sync + Debug + Clone {
    type StateType: State
    where
        <Self::StateType as State>::Time == ViewNumber;
    type Storage: Storage<Self::StateType> + Clone;
    type Networking: NetworkingImplementation<Message<Self::StateType, Self::SignatureKey>, Self::SignatureKey> + Clone;
    type SignatureKey: SignatureKey;
    type Election: Election<Self::SignatureKey, ViewNumber>
    where
        <Self::Election as Election<Self::SignatureKey, ViewNumber>>::StateType == Self::StateType;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type StateType: State
where
<Self::StateType as State>::Time == ViewNumber

State type for this consensus implementation

type Storage: Storage<Self::StateType> + Clone

Storage type for this consensus implementation

type Networking: NetworkingImplementation<Message<Self::StateType, Self::SignatureKey>, Self::SignatureKey> + Clone

Networking type for this consensus implementation

type SignatureKey: SignatureKey

The signature key type for this implementation

type Election: Election<Self::SignatureKey, ViewNumber>
where
<Self::Election as Election<Self::SignatureKey, ViewNumber>>::StateType == Self::StateType

Election Time is generic here to allow multiple implementations of election trait for difference consensus protocols