Trait hotshot::traits::NodeImplementation

pub trait NodeImplementation<TYPES>: Send + Sync + Debug + Clone + 'staticwhere
    TYPES: NodeType,{
    type Leaf: LeafType<NodeType = TYPES>;
    type Storage: Storage<TYPES, Self::Leaf> + Clone;
    type QuorumExchange: ConsensusExchange<TYPES, Message<TYPES, Self>>;
    type CommitteeExchange: ConsensusExchange<TYPES, Message<TYPES, Self>>;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type Leaf: LeafType<NodeType = TYPES>

Leaf type for this consensus implementation

type Storage: Storage<TYPES, Self::Leaf> + Clone

Storage type for this consensus implementation

type QuorumExchange: ConsensusExchange<TYPES, Message<TYPES, Self>>

Protocol for exchanging consensus proposals and votes.

type CommitteeExchange: ConsensusExchange<TYPES, Message<TYPES, Self>>

Protocol for exchanging data availability proposals and votes.