Trait hotshot::traits::State

pub trait State: Serialize + DeserializeOwned + Clone + Debug + Default + Hash + PartialEq<Self> + Eq + Send + Sync + Committable {
    type Error: Error + Debug + Send + Sync;
    type BlockType: Block;
    type Time: ConsensusTime;

    fn next_block(&self) -> Self::BlockType;
    fn validate_block(&self, block: &Self::BlockType, time: &Self::Time) -> bool;
    fn append(
        &self,
        block: &Self::BlockType,
        time: &Self::Time
    ) -> Result<Self, Self::Error>;
    fn on_commit(&self);
}

Abstraction over the state that blocks modify

This trait represents the behaviors that the ‘global’ ledger state must have:

A defined error type (Error)

The type of block that modifies this type of state (Block)

A method to get a template (empty) next block from the current state (next_block)

The ability to validate that a block is actually a valid extension of this state (validate_block)

The ability to produce a new state, with the modifications from the block applied (append)

Required Associated Types

type Error: Error + Debug + Send + Sync

The error type for this particular type of ledger state

type BlockType: Block

The type of block this state is associated with

type Time: ConsensusTime

Time abstraction needed for reward collection

Required Methods

fn next_block(&self) -> Self::BlockType

Returns an empty, template next block given this current state

fn validate_block(&self, block: &Self::BlockType, time: &Self::Time) -> bool

Returns true if and only if the provided block is valid and can extend this state

fn append(
&self,
block: &Self::BlockType,
time: &Self::Time
) -> Result<Self, Self::Error>

Appends the given block to this state, returning an new state

Errors

Should produce and error if appending this block would lead to an invalid state

fn on_commit(&self)

Gets called to notify the persistence backend that this state has been committed