Trait hotshot::traits::Storage

pub trait Storage<STATE>: Clone + Send + Sync where
    STATE: 'static + StateContents, {
    fn append<'life0, 'async_trait>(
        &'life0 self,
        views: Vec<ViewEntry<STATE>, Global>
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn cleanup_storage_up_to_view<'life0, 'async_trait>(
        &'life0 self,
        view: ViewNumber
    ) -> Pin<Box<dyn Future<Output = Result<usize, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_anchored_view<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<StoredView<STATE>, StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn commit<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn append_single_view<'life0, 'async_trait>(
        &'life0 self,
        view: StoredView<STATE>
    ) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Abstraction over on disk persistence of node state

This should be a cloneable handle to an underlying storage, with each clone pointing to the same underlying storage.

This trait has been constructed for object saftey over convenience.

Required Methods

fn append<'life0, 'async_trait>(
&'life0 self,
views: Vec<ViewEntry<STATE>, Global>
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Append the list of views to this storage

fn cleanup_storage_up_to_view<'life0, 'async_trait>(
&'life0 self,
view: ViewNumber
) -> Pin<Box<dyn Future<Output = Result<usize, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Cleans up the storage up to the given view. The given view number will still persist in this storage afterwards.

fn get_anchored_view<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<StoredView<STATE>, StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the latest anchored view

fn commit<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Commit this storage.

Provided Methods

fn append_single_view<'life0, 'async_trait>(
&'life0 self,
view: StoredView<STATE>
) -> Pin<Box<dyn Future<Output = Result<(), StorageError>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Insert a single view. Shorthand for

storage.append(vec![ViewEntry::Success(view)]).await