Trait hotshot::traits::TestableNodeImplementation

pub trait TestableNodeImplementation<TYPES>: NodeImplementation<TYPES>where
    TYPES: NodeType,{
    fn network_generator(
        expected_node_count: usize,
        num_bootstrap_nodes: usize
    ) -> Box<dyn Fn(u64) -> <<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK + 'static, Global>;
    fn committee_generator(
    ) -> Box<dyn Fn(Arc<<<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <Self::CommitteeExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>;
    fn quorum_generator(
    ) -> Box<dyn Fn(Arc<<<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>;
    fn state_create_random_transaction(
        state: Option<&<TYPES as NodeType>::StateType>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Self::Leaf,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <<TYPES as NodeType>::BlockType as Block>::Transaction;
    fn block_genesis() -> <TYPES as NodeType>::BlockType;
    fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn generate_test_key(
        id: u64
    ) -> <<TYPES as NodeType>::SignatureKey as SignatureKey>::PrivateKey;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Methods

fn network_generator(
expected_node_count: usize,
num_bootstrap_nodes: usize
) -> Box<dyn Fn(u64) -> <<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK + 'static, Global>

generates a network given an expected node count

fn committee_generator(
) -> Box<dyn Fn(Arc<<<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <Self::CommitteeExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>

generates a committee communication channel given the network

fn quorum_generator(
) -> Box<dyn Fn(Arc<<<Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking as CommunicationChannel<TYPES, Message<TYPES, Self>, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>>::NETWORK>) -> <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static, Global>

generates a quorum communication channel given the network

fn state_create_random_transaction(
state: Option<&<TYPES as NodeType>::StateType>,
rng: &mut dyn RngCore,
padding: u64
) -> <<TYPES as NodeType>::BlockType as Block>::Transaction

Get the number of messages in-flight from quorum exchange.

Some implementations will not be able to tell how many messages there are in-flight. These implementations should return None. Get the number of messages in-flight from committee exchange.

Some implementations will not be able to tell how many messages there are in-flight. These implementations should return None. Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction(
leaf: &Self::Leaf,
rng: &mut dyn RngCore,
padding: u64
) -> <<TYPES as NodeType>::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> <TYPES as NodeType>::BlockType

generate a genesis block

fn txn_count(block: &<TYPES as NodeType>::BlockType) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>(
storage: &'life0 Self::Storage
) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait, Global>>where
'life0: 'async_trait,
Self: 'async_trait,

Return the full internal state. This is useful for debugging.

fn generate_test_key(
id: u64
) -> <<TYPES as NodeType>::SignatureKey as SignatureKey>::PrivateKey

The private key of the node id in a test.