hotshot

Module types

Contains types used by the crate

Modules

event 🔒

Events that a SystemContext instance can emit

handle 🔒

Provides an event-streaming handle for a SystemContext running in the background

Structs

Event

A status event emitted by a HotShot instance

Message

Incoming message

SystemContextHandle

Event streaming handle for a SystemContext instance running in the background

Enums

EventType

The type and contents of a status event emitted by a HotShot instance

Traits

SignatureKey

Trait for abstracting public key signatures Self is the public key type

Type Aliases

BLSPrivKey

BLS private key used to sign a message

BLSPubKey

BLS public key used to verify a signature

SchnorrPrivKey

SchnorrPubKey